# api_gateway/internal/app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/carts": {
            "get": {
                "description": "Получает содержимое корзины",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Get cart details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contents",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.Cart"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет корзину",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Update cart contents",
                "parameters": [
                    {
                        "description": "Update payload",
                        "name": "update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.UpdateCartReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.Cart"
                        }
                    },
                    "403": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Очищает корзину",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Delete a user's cart",
                "parameters": [
                    {
                        "description": "User payload",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.GetCartReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.GeneralOrderResponse"
                        }
                    },
                    "403": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/carts/add/product": {
            "post": {
                "description": "Добавляет товары в корзину",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Add products to cart",
                "parameters": [
                    {
                        "description": "Items payload",
                        "name": "items",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.AddProducts2Cart"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.GeneralOrderResponse"
                        }
                    },
                    "403": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/carts/product": {
            "delete": {
                "description": "Удаляет указанные товары из корзины",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Remove specific products from cart",
                "parameters": [
                    {
                        "description": "Items payload",
                        "name": "items",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.DeleteProductsfromCartReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.Cart"
                        }
                    },
                    "403": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders/completed": {
            "put": {
                "description": "Помечает заказ как завершённый",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Mark an order as completed",
                "parameters": [
                    {
                        "description": "Update payload",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.UpdateOrderReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.Order"
                        }
                    },
                    "403": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders/create": {
            "post": {
                "description": "Создаёт новый заказ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "parameters": [
                    {
                        "description": "Order payload",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.CreateOrderReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.Order"
                        }
                    },
                    "403": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders/getall": {
            "get": {
                "description": "Возвращает список заказов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Retrieve all orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by customer ID",
                        "name": "userId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.GetAllOrdersRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.AddProducts2Cart": {
            "type": "object",
            "properties": {
                "productId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.Cart": {
            "type": "object",
            "properties": {
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.ProductOrder"
                    }
                },
                "totalPrice": {
                    "type": "number"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.CreateOrderReq": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "contactNumber": {
                    "type": "string"
                },
                "cordination": {
                    "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.Location"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.DeleteProductsfromCartReq": {
            "type": "object",
            "properties": {
                "productId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
        "github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.GeneralOrderResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.GetAllOrdersRes": {
            "type": "object",
            "properties": {
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.Order"
                    }
                }
            }
        },
        "github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.GetCartReq": {
            "type": "object",
            "properties": {
                "userId": {
                    "type": "string"
                }
            }
        },
        "github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.Location": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "string"
                },
                "longtitude": {
                    "type": "string"
                }
            }
        },
        "github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.Order": {
            "type": "object",
            "properties": {
                "cartId": {
                    "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.Cart"
                },
                "comment": {
                    "type": "string"
                },
                "contactNumber": {
                    "type": "string"
                },
                "cordination": {
                    "$ref": "#/definitions/github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.Location"
                },
                "createdat": {
                    "type": "string"
                },
                "orderId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedat": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.ProductOrder": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "number"
                },
                "productId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.UpdateCartReq": {
            "type": "object",
            "properties": {
                "productId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "github_com_diyorbek_E-Commerce_BOT_api_gateway_internal_entity_order.UpdateOrderReq": {
            "type": "object",
            "properties": {
                "orderId": {
                    "type": "string"
                }
            }
        },
...
    }
}